完善后台管理&前台主页

+++ TPro.EntityFramework/Entity/Jurisdiction.cs
@@ -2,9 +2,9 @@
 {
     public class Jurisdiction
     {
-        public int Id { get; set; }
-        public int RoleId { get; set; }
-        public int RouteId { get; set; }
+        public long Id { get; set; }
+        public long RoleId { get; set; }
+        public long RouteId { get; set; }
         public bool IsAllowed { get; set; }
         public virtual TPRole Role { get; set; }
         public virtual RoutePath Route { get; set; }
