+++ TPro.EntityFramework/Entity/UserRole.cs
@@ -6,13 +6,13 @@ namespace TPro.EntityFramework.Entity
     public class UserRole
     {
         [Key]
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         [ForeignKey(nameof(User))]
-        public int UserId { get; set; }
+        public long UserId { get; set; }
 
         [ForeignKey(nameof(Role))]
-        public int RoleId { get; set; }
+        public long RoleId { get; set; }
 
         public string Name { get; set; } = "";
         public virtual TPUser User { get; set; }
